refactor!: rename `etl::clock` to `synfig::clock`
Moved ETL benchmark test to synfig-core due to `clock` requirement

BREAKING CHANGE: both ETL and synfig API changed:
- `etl::clock` is now `synfig::clock`
