+++ synfig-core/src/synfig/clock.h
@@ -1,7 +1,6 @@
-// <clock> -*- C++ -*-
-/*! ========================================================================
-** Extended Template and Library
-** Clock Abstraction
+/* === S Y N F I G ========================================================= */
+/*!	\file clock.h
+**	\brief A timer class
 **
 ** Copyright (c) 2002 Robert B. Quattlebaum Jr.
 **
@@ -24,14 +23,18 @@
 
 /* === S T A R T =========================================================== */
 
-#ifndef __ETL__CLOCK__
-#define __ETL__CLOCK__
+#ifndef SYNFIG_CLOCK_H
+#define SYNFIG_CLOCK_H
 
 /* === H E A D E R S ======================================================= */
 
 #include <chrono>
 
-namespace etl {
+/* === T Y P E D E F S ===================================================== */
+
+/* === C L A S S E S & S T R U C T S ======================================= */
+
+namespace synfig {
 
 	class clock {
 		std::chrono::time_point<std::chrono::steady_clock> base_time_;
@@ -44,12 +47,10 @@ namespace etl {
 		void reset() {
 			base_time_ = std::chrono::steady_clock::now();
 		}
-		//{ get_current_time(base_time); }
 
 		float operator()() const {
 			return std::chrono::duration<float>(std::chrono::steady_clock::now() - base_time_).count();
 		}
-		//{ return this->timestamp_to_seconds(get_current_time()-base_time); }
 
 		float pop_time() {
 			// Grab the old base time
@@ -62,8 +63,6 @@ namespace etl {
 		}
 	};
 }
-
-//using etl::clock;
 
 /* === E N D =============================================================== */
