+++ synfig-core/test/benchmark.cpp
@@ -1,7 +1,8 @@
-/*! ========================================================================
-** Extended Template and Library Test Suite
-** Hermite Curve Test
+/* === S Y N F I G ========================================================= */
+/*!	\file benchmark.cpp
+**	\brief Benchmark test
 **
+**	\legal
 ** Copyright (c) 2002 Robert B. Quattlebaum Jr.
 **
 ** This file is part of Synfig.
@@ -23,13 +24,15 @@
 
 /* === H E A D E R S ======================================================= */
 
-#include <ETL/clock>
+#include <cstdio>
+
 #include <ETL/hermite>
 #include <ETL/angle>
 #include <ETL/surface>
 #include <ETL/gaussian>
 #include <ETL/calculus>
-#include <stdio.h>
+
+#include <synfig/clock.h>
 
 /* === M A C R O S ========================================================= */
 
@@ -126,7 +129,7 @@ void angle_atan2_speed_test(void)
 int surface_and_gaussian_blur_test()
 {
 	int ret=0;
-	etl::clock MyTimer;
+	synfig::clock MyTimer;
 	float endtime;
 
 	{
@@ -157,7 +160,7 @@ int hermite_int_test()
 	hermite<int>::time_type f;
 	int i;
 
-	etl::clock timer;
+	synfig::clock timer;
 	float t;
 
 	Hermie.p1()=0;
@@ -196,7 +199,7 @@ int hermite_float_test(void)
 	float f; int i;
 
 	hermite<float> Hermie;
-	etl::clock timer;
+	synfig::clock timer;
 	double t;
 
 	Hermie.p1()=0;
@@ -235,7 +238,7 @@ int hermite_double_test(void)
 	float f;
 
 	hermite<double> Hermie;
-	etl::clock timer;
+	synfig::clock timer;
 	double t;
 
 	Hermie.p1()=0;
@@ -272,7 +275,7 @@ int hermite_angle_test(void)
 	float f;
 
 	hermite<angle> Hermie;
-	etl::clock timer;
+	synfig::clock timer;
 	angle tmp;
 	double t;
