+++ synfig-studio/src/gui/states/state_width.cpp
@@ -34,8 +34,6 @@
 #	include <config.h>
 #endif
 
-#include <ETL/clock>
-
 #include <gui/app.h>
 #include <gui/canvasview.h>
 #include <gui/docks/dock_toolbox.h>
@@ -49,6 +47,7 @@
 #include <gui/workarea.h>
 
 #include <synfig/blinepoint.h>
+#include <synfig/clock.h>
 #include <synfig/general.h>
 #include <synfig/valuenodes/valuenode_wplist.h>
 
@@ -86,8 +85,7 @@ class studio::StateWidth_Context : public sigc::trackable
 
 	std::map<handle<Duck>,Real>	changetable;
 
-	etl::clock	clocktime;
-	// Real		lastt; // unused
+	synfig::clock	clocktime;
 
 	bool added;
